+++ src/staff-list.js
@@ -7,7 +7,7 @@
 registerPlugin(
     {
         name: 'Staff List',
-        version: '1.1.0',
+        version: '1.2.0',
         description:
             'With this script, the bot will automatically keep track of the online status of predefined staff members and post it to a chosen channel description.',
         author: 'RLNT',
@@ -38,27 +38,122 @@ registerPlugin(
             },
             {
                 name: 'clickable',
-                title: 'Will make the usernames in the list clickable (hyperlink).',
+                title: 'Do you want the usernames in the list to be clickable (hyperlinks)?',
                 type: 'select',
                 options: [ 'Hyperlink (clickable)', 'Plain Text' ]
+            },
+            {
+                name: 'template',
+                title: 'Do you want to use a custom template to format your staff list? (*) | Advanced Option',
+                type: 'select',
+                options: [ 'Yes', 'No' ]
+            },
+            {
+                name: 'tUsername',
+                title:
+                    'Define what the username in the list should look like. | placeholders: %name% - nickname of the user',
+                type: 'string',
+                placeholder: '[B]%name%[/B]',
+                indent: 1,
+                conditions: [
+                    {
+                        field: 'template',
+                        value: 0
+                    }
+                ]
+            },
+            {
+                name: 'tPhraseOnline',
+                title: 'Define what the phrase if a user is online should look like.',
+                type: 'string',
+                placeholder: '[COLOR=#00ff00][B]ONLINE[/B][/COLOR]',
+                indent: 1,
+                conditions: [
+                    {
+                        field: 'template',
+                        value: 0
+                    }
+                ]
+            },
+            {
+                name: 'tPhraseOffline',
+                title: 'Define what the phrase if a user is offline should look like.',
+                type: 'string',
+                placeholder: '[COLOR=#ff0000][B]OFFLINE[/B][/COLOR]',
+                indent: 1,
+                conditions: [
+                    {
+                        field: 'template',
+                        value: 0
+                    }
+                ]
+            },
+            {
+                name: 'tMemberLine',
+                title:
+                    'Define what a full line in the member list should look like. | placeholders: %name% - formatted username, %status% - formatted online status, %lb% - line break',
+                type: 'multiline',
+                placeholder: '%name% [COLOR=#aaff00]>[/COLOR] %status%',
+                indent: 1,
+                conditions: [
+                    {
+                        field: 'template',
+                        value: 0
+                    }
+                ]
+            },
+            {
+                name: 'tGroupSection',
+                title:
+                    'Define what a group section should look like. | placeholders: %group% - formatted group name, %members% - formatted member lines, %lb% - line break',
+                type: 'multiline',
+                placeholder: '[center]%group%\n%members%\n____________________\n[/center]\n%lb%',
+                indent: 1,
+                conditions: [
+                    {
+                        field: 'template',
+                        value: 0
+                    }
+                ]
             },
             {
                 name: 'separator',
                 title: 'The separator that should separate the groups in the channel.',
                 type: 'multiline',
-                placeholder: '_______________________________________'
+                placeholder: '_______________________________________',
+                indent: 1,
+                conditions: [
+                    {
+                        field: 'template',
+                        value: 1
+                    }
+                ]
             },
             {
                 name: 'phraseOnline',
                 title: 'The phrase that should be shown if the corresponding user is online.',
                 type: 'string',
-                placeholder: '[COLOR=#00ff00][B]ONLINE[/B][/COLOR]'
+                placeholder: '[COLOR=#00ff00][B]ONLINE[/B][/COLOR]',
+                indent: 1,
+                conditions: [
+                    {
+                        field: 'template',
+                        value: 1
+                    }
+                ]
             },
             {
                 name: 'phraseOffline',
                 title: 'The phrase that should be shown if the corresponding user is offline.',
                 type: 'string',
-                placeholder: '[COLOR=#ff0000][B]OFFLINE[/B][/COLOR]'
+                placeholder: '[COLOR=#ff0000][B]OFFLINE[/B][/COLOR]',
+                indent: 1,
+                conditions: [
+                    {
+                        field: 'template',
+                        value: 1
+                    }
+                ]
             },
             {
                 name: 'staffGroups',
@@ -106,6 +201,7 @@ registerPlugin(
         // GLOBAL VARS
         const prefix = 'Staff-List';
         let memberList = [];
+        let groupList = [];
 
         // GLOBAL FUNCTIONS
         function log(message) {
@@ -128,6 +224,9 @@ registerPlugin(
             } else if (config.staffGroups === undefined || config.staffGroups.length === 0) {
                 log('There are no staff groups selected to be displayed in the staff list. Deactivating script...');
                 return;
+            } else if (config.template === undefined) {
+                log('There was no formatting option selected. Deactivating script...');
+                return;
             } else {
                 log('The script has loaded successfully!');
                 waitForBackend();
@@ -140,31 +239,28 @@ registerPlugin(
 
             config.staffGroups.forEach(group => {
                 if (group.id === undefined || backend.getServerGroupByID(group.id) === undefined) return;
-                if (group.clients === undefined || group.clients.length === 0) {
-                    if (group.groups === undefined || group.groups.length === 0) {
-                        group.clients = [];
-                        group.groups = [ group.id ];
-                    } else {
-                        group.clients = [];
-                        group.groups.map(id => backend.getServerGroupByID(id) !== undefined && id !== group.id);
-                        group.groups.concat([ group.id ]);
-                    }
+                if (group.clients === undefined || group.clients.length === 0) group.clients = [];
+                if (group.groups === undefined || group.groups.length === 0) {
+                    group.groups = [ group.id ];
                 } else {
-                    if (group.groups === undefined || group.groups.length === 0) {
-                        group.groups = [ group.id ];
-                    } else {
-                        group.groups.map(id => backend.getServerGroupByID(id) !== undefined && id !== group.id);
-                        group.groups.concat([ group.id ]);
-                    }
+                    group.groups.map(id => backend.getServerGroupByID(id) !== undefined && id !== group.id);
+                    group.groups.push(group.id);
                 }
                 if (group.name === undefined || group.name === '') {
                     group.name = '[size=12][B]' + backend.getServerGroupByID(group.id).name() + '[/B][/size]';
                 }
 
+                groupList = groupList.concat(group.groups);
                 staffGroups.push(group);
             });
 
             return staffGroups;
+        }
+
+        function validateDatabase() {
+            store.getKeys().forEach(key => {
+                if (!groupList.includes(store.get(key)[1])) removeMember(key);
+            });
         }
 
         function updateMemberList() {
@@ -188,6 +284,13 @@ registerPlugin(
                 store.set(uid, [ nick, group ]);
             }
             updateMemberList();
+        }
+
+        function removeMember(uid) {
+            if (store.getKeys().includes(uid)) {
+                store.unset(uid);
+                updateMemberList();
+            }
         }
 
         function getClientGroup(client, staffGroups) {
@@ -243,11 +346,28 @@ registerPlugin(
             return [ membersOnline, membersOffline ];
         }
 
-        function updateDescription(staffGroups, clickable, phraseOnline, phraseOffline, separator, channel) {
-            let description = '';
+        function updateDescription(staffGroups, channel) {
+            const template = config.template == 0;
+            const clickable = config.clickable == 0;
+            let username, memberLine, groupSection, separator, phraseOnline, phraseOffline;
+
+            if (config.template == 0) {
+                username = config.tUsername || '[B]%name%[/B]';
+                memberLine = config.tMemberLine || '%name% [COLOR=#aaff00]>[/COLOR] %status%';
+                groupSection = config.tGroupSection || '[center]%group%\n%members%____________________[/center]';
+                phraseOnline = config.tPhraseOnline || '[COLOR=#00ff00][B]ONLINE[/B][/COLOR]';
+                phraseOffline = config.tPhraseOffline || '[COLOR=#ff0000][B]OFFLINE[/B][/COLOR]';
+            } else {
+                separator = config.separator || '_______________________________________';
+                phraseOnline = config.phraseOnline || '[COLOR=#00ff00][B]ONLINE[/B][/COLOR]';
+                phraseOffline = config.phraseOffline || '[COLOR=#ff0000][B]OFFLINE[/B][/COLOR]';
+            }
+
             const sortedMemberList = getSortedMemberList();
             const membersOnline = sortedMemberList[0];
             const membersOffline = sortedMemberList[1];
+            let description = '';
+
             staffGroups.forEach(staffGroup => {
                 let membersToList = '';
                 membersOnline.forEach(member => {
@@ -256,11 +376,24 @@ registerPlugin(
                     const memberGroup = member[2];
 
                     if (staffGroup.id === memberGroup) {
+                        let memberName = '';
                         if (clickable) {
-                            membersToList += `[URL=client://0/${memberUid}]${memberNick}[/URL] - ${phraseOnline}\n`;
+                            memberName = `[URL=client://0/${memberUid}]${memberNick}[/URL]`;
                         } else {
-                            membersToList += `${memberNick} - ${phraseOnline}\n`;
+                            memberName = `${memberNick}`;
                         }
+
+                        let memberToList = '';
+                        if (template) {
+                            memberToList = memberLine
+                                .replace('%name%', username.replace('%name%', memberName))
+                                .replace('%status%', phraseOnline)
+                                .replace('%lb%', '\n');
+                        } else {
+                            memberToList = `${memberName} - ${phraseOnline}`;
+                        }
+
+                        membersToList += `${memberToList}\n`;
                     }
                 });
                 membersOffline.forEach(member => {
@@ -269,15 +402,37 @@ registerPlugin(
                     const memberGroup = member[2];
 
                     if (staffGroup.id === memberGroup) {
+                        let memberName = '';
                         if (clickable) {
-                            membersToList += `[URL=client://0/${memberUid}]${memberNick}[/URL] - ${phraseOffline}\n`;
+                            memberName = `[URL=client://0/${memberUid}]${memberNick}[/URL]`;
                         } else {
-                            membersToList += `${memberNick} - ${phraseOffline}\n`;
+                            memberName = `${memberNick}`;
                         }
+
+                        let memberToList = '';
+                        if (template) {
+                            memberToList = memberLine
+                                .replace('%name%', username.replace('%name%', memberName))
+                                .replace('%status%', phraseOffline)
+                                .replace('%lb%', '\n');
+                        } else {
+                            memberToList = `${memberName} - ${phraseOffline}`;
+                        }
+
+                        membersToList += `${memberToList}\n`;
                     }
                 });
 
-                if (membersToList !== '') description += `${staffGroup.name}\n${membersToList}${separator}\n`;
+                if (membersToList !== '') {
+                    if (template) {
+                        description += groupSection
+                            .replace('%group%', staffGroup.name)
+                            .replace('%members%', membersToList.substring(0, membersToList.length - 1))
+                            .replace('%lb%', '\n');
+                    } else {
+                        description += `${staffGroup.name}\n${membersToList}${separator}\n`;
+                    }
+                }
             });
 
             // set new description to channel
@@ -288,23 +443,26 @@ registerPlugin(
         function main() {
             // VARIABLES
             const staffGroups = validateListGroups();
-            const separator = config.separator || '_______________________________________';
-            const clickable = config.clickable == 0 || true;
             const channel = backend.getChannelByID(config.channel);
-            const phraseOnline = config.phraseOnline || '[COLOR=#00ff00][B]ONLINE[/B][/COLOR]';
-            const phraseOffline = config.phraseOffline || '[COLOR=#ff0000][B]OFFLINE[/B][/COLOR]';
+
+            // validate database
+            validateDatabase();
 
             // store all online listed members
             backend.getClients().forEach(client => {
                 const staffGroup = getClientGroup(client, staffGroups);
-                if (staffGroup !== undefined) store.set(client.uid(), [ client.nick(), staffGroup.id ]);
+                if (staffGroup !== undefined) {
+                    storeMember(client.uid(), client.nick(), staffGroup.id);
+                } else {
+                    removeMember(client.uid());
+                }
             });
 
             // update the cached member list
             updateMemberList();
 
             // update the description for all currently known users
-            updateDescription(staffGroups, clickable, phraseOnline, phraseOffline, separator, channel);
+            updateDescription(staffGroups, channel);
 
             // MOVE EVENT
             event.on('clientMove', event => {
@@ -317,15 +475,46 @@ registerPlugin(
                 const group = getClientGroup(client, staffGroups);
 
                 // make sure it's a user that has to be listed
-                if (group !== null) {
+                if (group !== undefined) {
                     // on connect or disconnect
                     if (fromChannel === undefined || toChannel === undefined) {
                         // make sure user is stored
                         storeMember(uid, nick, group.id);
 
                         // update the description
-                        updateDescription(staffGroups, clickable, phraseOnline, phraseOffline, separator, channel);
+                        updateDescription(staffGroups, channel);
                     }
+                }
+            });
+
+            // SERVER GROUP ADDED EVENT
+            event.on('serverGroupAdded', event => {
+                const client = event.client;
+                if (client.isSelf()) return;
+                const eventGroup = event.serverGroup.id();
+                const clientGroup = getClientGroup(client, staffGroups);
+
+                if (groupList.includes(eventGroup)) {
+                    storeMember(client.uid(), client.nick(), clientGroup.id);
+                    updateDescription(staffGroups, channel);
+                }
+            });
+
+            // SERVER GROUP REMOVE EVENT
+            event.on('serverGroupRemoved', event => {
+                const client = event.client;
+                if (client.isSelf()) return;
+                const eventGroup = event.serverGroup.id();
+                const clientGroup = getClientGroup(client, staffGroups);
+
+                if (groupList.includes(eventGroup)) {
+                    if (clientGroup === undefined) {
+                        removeMember(client.uid());
+                    } else {
+                        storeMember(client.uid(), client.nick(), clientGroup.id);
+                    }
+
+                    updateDescription(staffGroups, channel);
                 }
             });
         }